Fixed Shipping and Payment Provider and Total order sum. And cleand up files
