+++ src/components/checkout-cards/shipping.tsx
@@ -1,26 +1,29 @@
-import { FC, useState, useContext } from "react" 
-import { Shipping, shippingList } from "../../data/shippingList" 
+import { FC, useContext } from "react" 
+import { shippingList } from "../../data/shippingList" 
 import Radio from '@mui/material/Radio'
 import RadioGroup from '@mui/material/RadioGroup'
 import FormControlLabel from '@mui/material/FormControlLabel'
 import FormControl from '@mui/material/FormControl'
-import { ShippingContext } from "../context/checkout-context/shipping-context"
+import { ShippingContext } from "../context/shipping-provider"
 import FormLabel from '@mui/material/FormLabel'
-
+import { smallImageStyle } from "../../css/common"
 
 
 interface Props {}
 
 const ShippingCard: FC<Props> = (props) => {
 
-    const { shippingState, setShippingState } = useContext(ShippingContext) 
-    
+
+    const { setShippingState, shippingState } = useContext(ShippingContext)
+
+
     const getShippingDate = (shippingOption: number | undefined) => {
         let actualDate: Date = new Date()
         actualDate = new Date(new Date().setHours(new Date().getHours() + shippingOption!))
         let newShippingDate = actualDate.toLocaleDateString()
-        return newShippingDate
-    }
+            return newShippingDate
+
+        }
 
     return (
 
@@ -36,30 +39,33 @@ const ShippingCard: FC<Props> = (props) => {
 
                  {shippingList.map((shippingOption) => {
                     return (
-                        <tr key={shippingOption.id}>
+                
                     <>
+                    <tr key={shippingOption.id}>
+                
                     <FormControlLabel 
                         value={shippingOption.id} control={<Radio />} label={shippingOption.title}
-                        onChange={() => {setShippingState(shippingOption)}}
-                        /> 
-        
+                            onChange={() => 
+                                {setShippingState(shippingOption)}} 
+                            />
+                        
+                        <img style={smallImageStyle} src={shippingOption.image}/>
                         <p>{shippingOption.price} kr</p>
-                        <p>{
-                            shippingState && (shippingState.id == shippingOption.id)?
-                            <h2>{getShippingDate(shippingOption.shippingDate)}</h2>:undefined
+
+                        <p>{shippingState && (shippingState.id == shippingOption.id) ?
+                            <p>Beräknad leverans: {getShippingDate(shippingOption.shippingDate)}</p> : undefined
                             }</p>
-                     
-                  
-                    </>
+
                     </tr>
+                    
+
+                    </>
                 )
             })
         }
-        </>
-               
+        </>    
             </RadioGroup>
         </FormControl>
-
     )
 }
