+++ src/components/content.tsx
@@ -3,7 +3,7 @@ import { Route, Routes } from "react-router-dom";
 import Home from "./pages/home"
 import Products from "./pages/product"
 import CheckOut from "./pages/checkout"
-import ProductDetail from "../components/pages-cards/productDetail"
+import ProductDetail from "./page-cards/product-detail"
 
 interface Props {}
