+++ src/components/pages/home.tsx
@@ -2,7 +2,7 @@ import React from "react"
 import { FC } from "react"
 import { productList } from '../../data/productlist' 
 import { imgContainer} from "../../css/common"
-import ProductCard from "../pages-cards/productCard"
+import ProductCard from "../page-cards/product-card"
 
 
 interface Props {}
